requirements.txt
streamlit
groq
python-docx
pdfminer.six
PyPDF2
python-dotenv
pdfplumber
pillow 
pytesseract 
pdf2image